move website type tests
