+++ backend/search_module/scraper/search/module_test.go
@@ -1,7 +1,6 @@
 package search_test
 
 import (
-	"encoding/json"
 	"search_module/scraper"
 	"search_module/scraper/search"
 	"testing"
@@ -30,34 +29,6 @@ func (tws testWebsiteSearch) GetResults(phrase string, page int) (search.Result,
 		},
 	}
 	return sr, nil
-}
-
-func TestUnmarshallingWebsiteTypeWithCorrectValueShouldReturnEnumWithAppropriateType(t *testing.T) {
-	jsonInput := []byte(`"ceneo"`)
-	var wt scraper.WebsiteType
-
-	err := json.Unmarshal(jsonInput, &wt)
-
-	assert.Nil(t, err)
-	assert.Equal(t, wt, scraper.Ceneo)
-}
-
-func TestUnmarshallingWebsiteTypeWithEmptyValueShouldReturnError(t *testing.T) {
-	jsonInput := []byte(`""`)
-	var wt scraper.WebsiteType
-
-	err := json.Unmarshal(jsonInput, &wt)
-
-	assert.NotNil(t, err)
-}
-
-func TestUnmarshallingWebsiteTypeWithNonExistingValueShouldReturnError(t *testing.T) {
-	jsonInput := []byte(`"not_exist"`)
-	var wt scraper.WebsiteType
-
-	err := json.Unmarshal(jsonInput, &wt)
-
-	assert.NotNil(t, err)
 }
 
 func TestSearchShouldReturnResultsFromWebsiteSearchImplementation(t *testing.T) {
